Nessesary for postgres connection
This lib folder with these Jar files MUST be added into the project folders setting in order to properly connect to the postgres server. Failure to do so will cause an error. *Also the information regarding the port and password etc. must match those of the server where the database is stored for successful connection*.